Move launcher code into the main package
This will allow us to define setuptools entry points
