+++ quodlibet/po/POTFILES.in
@@ -2,8 +2,6 @@ data/exfalso.appdata.xml.in
 data/exfalso.desktop.in
 data/quodlibet.appdata.xml.in
 data/quodlibet.desktop.in
-exfalso.py
-operon.py
 quodlibet/browsers/albums/main.py
 quodlibet/browsers/albums/prefs.py
 quodlibet/browsers/audiofeeds.py
@@ -30,6 +28,7 @@ quodlibet/devices/_base.py
 quodlibet/devices/__init__.py
 quodlibet/devices/ipod.py
 quodlibet/devices/storage.py
+quodlibet/exfalso.py
 quodlibet/ext/covers/artwork_url.py
 quodlibet/ext/covers/lastfm.py
 quodlibet/ext/covers/musicbrainz.py
@@ -73,8 +72,8 @@ quodlibet/ext/events/trayicon/menu.py
 quodlibet/ext/events/trayicon/prefs.py
 quodlibet/ext/events/trayicon/systemtray.py
 quodlibet/ext/events/viewlyrics.py
-quodlibet/ext/events/write_cover.py
 quodlibet/ext/events/waveformseekbar.py
+quodlibet/ext/events/write_cover.py
 quodlibet/ext/events/zeitgeist_client.py
 quodlibet/ext/gstreamer/compressor.py
 quodlibet/ext/gstreamer/crossfeed.py
@@ -131,6 +130,7 @@ quodlibet/formats/__init__.py
 quodlibet/__init__.py
 quodlibet/library/__init__.py
 quodlibet/library/libraries.py
+quodlibet/main.py
 quodlibet/operon/base.py
 quodlibet/operon/commands.py
 quodlibet/operon/main.py
