+++ quodlibet/quodlibet/__init__.py
@@ -21,13 +21,13 @@ from .util.i18n import _, C_, N_, ngettext, npgettext
 from .util.dprint import print_d, print_e, print_w
 from ._init import init_cli, init
 from ._main import get_base_dir, is_release, get_user_dir, app, \
-    set_application_info, init_plugins, enable_periodic_save, main, \
+    set_application_info, init_plugins, enable_periodic_save, run, \
     finish_first_session, get_image_dir, is_first_session, \
     get_build_description, get_build_version
 
 
 _, C_, N_, ngettext, npgettext, is_release, init, init_cli, print_e, \
     get_base_dir, print_w, print_d, get_user_dir, app, set_application_info, \
-    init_plugins, enable_periodic_save, main, finish_first_session, \
+    init_plugins, enable_periodic_save, run, finish_first_session, \
     get_image_dir, is_first_session, get_build_description, \
     get_build_version
